Removed code for custom sign-models
